+++ packages/node/src/indexer/fetch.service.spec.ts
@@ -1,4 +1,4 @@
-// Copyright 2020-2023 SubQuery Pte Ltd authors & contributors
+// Copyright 2020-2024 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
 import {
@@ -22,7 +22,6 @@ const mockTempDs: EthereumProjectDsTemplate[] = [
     kind: EthereumDatasourceKind.Runtime,
     assets: new Map(),
     mapping: {
-      entryScript: '',
       file: '',
       handlers: [
         {
@@ -40,7 +39,6 @@ const mockTempDs: EthereumProjectDsTemplate[] = [
     kind: EthereumDatasourceKind.Runtime,
     assets: new Map(),
     mapping: {
-      entryScript: '',
       file: '',
       handlers: [
         {