+++ packages/node/src/indexer/fetch.service.ts
@@ -1,4 +1,4 @@
-// Copyright 2020-2023 SubQuery Pte Ltd authors & contributors
+// Copyright 2020-2024 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
 import { Inject, Injectable } from '@nestjs/common';
@@ -30,6 +30,189 @@ import {
 const BLOCK_TIME_VARIANCE = 5000;
 
 const INTERVAL_PERCENT = 0.9;
+
+function appendDsOptions(
+  dsOptions: SubqlEthereumProcessorOptions | SubqlEthereumProcessorOptions[],
+  conditions: DictionaryQueryCondition[],
+): void {
+  const queryAddressLimit = yargsOptions.argv['query-address-limit'];
+  if (Array.isArray(dsOptions)) {
+    const addresses = dsOptions.map((option) => option.address).filter(Boolean);
+
+    if (addresses.length > queryAddressLimit) {
+      logger.debug(
+        `Addresses length: ${addresses.length} is exceeding limit: ${queryAddressLimit}. Consider increasing this value with the flag --query-address-limit  `,
+      );
+    }
+
+    if (addresses.length !== 0 && addresses.length <= queryAddressLimit) {
+      conditions.push({
+        field: 'address',
+        value: addresses,
+        matcher: 'in',
+      });
+    }
+  } else {
+    if (dsOptions?.address) {
+      conditions.push({
+        field: 'address',
+        value: dsOptions.address.toLowerCase(),
+        matcher: 'equalTo',
+      });
+    }
+  }
+}
+
+function eventFilterToQueryEntry(
+  filter: EthereumLogFilter,
+  dsOptions: SubqlEthereumProcessorOptions | SubqlEthereumProcessorOptions[],
+): DictionaryQueryEntry {
+  const conditions: DictionaryQueryCondition[] = [];
+  appendDsOptions(dsOptions, conditions);
+  if (filter.topics) {
+    for (let i = 0; i < Math.min(filter.topics.length, 4); i++) {
+      const topic = filter.topics[i];
+      if (!topic) {
+        continue;
+      }
+      const field = `topics${i}`;
+
+      if (topic === '!null') {
+        conditions.push({
+          field,
+          value: false as any, // TODO update types to allow boolean
+          matcher: 'isNull',
+        });
+      } else {
+        conditions.push({
+          field,
+          value: eventToTopic(topic),
+          matcher: 'equalTo',
+        });
+      }
+    }
+  }
+  return {
+    entity: 'evmLogs',
+    conditions,
+  };
+}
+
+function callFilterToQueryEntry(
+  filter: EthereumTransactionFilter,
+  dsOptions: SubqlEthereumProcessorOptions | SubqlEthereumProcessorOptions[],
+): DictionaryQueryEntry {
+  const conditions: DictionaryQueryCondition[] = [];
+  appendDsOptions(dsOptions, conditions);
+
+  for (const condition of conditions) {
+    if (condition.field === 'address') {
+      condition.field = 'to';
+    }
+  }
+  if (filter.from) {
+    conditions.push({
+      field: 'from',
+      value: filter.from.toLowerCase(),
+      matcher: 'equalTo',
+    });
+  }
+  const optionsAddresses = conditions.find((c) => c.field === 'to');
+  if (!optionsAddresses) {
+    if (filter.to) {
+      conditions.push({
+        field: 'to',
+        value: filter.to.toLowerCase(),
+        matcher: 'equalTo',
+      });
+    } else if (filter.to === null) {
+      conditions.push({
+        field: 'to',
+        value: true as any, // TODO update types to allow boolean
+        matcher: 'isNull',
+      });
+    }
+  } else if (optionsAddresses && (filter.to || filter.to === null)) {
+    logger.warn(
+      `TransactionFilter 'to' conflict with 'address' in data source options`,
+    );
+  }
+
+  if (filter.function === null || filter.function === '0x') {
+    conditions.push({
+      field: 'func',
+      value: true,
+      matcher: 'isNull',
+    });
+  } else if (filter.function) {
+    conditions.push({
+      field: 'func',
+      value: functionToSighash(filter.function),
+      matcher: 'equalTo',
+    });
+  }
+
+  return {
+    entity: 'evmTransactions',
+    conditions,
+  };
+}
+
+type GroupedEthereumProjectDs = SubqlDatasource & {
+  groupedOptions?: SubqlEthereumProcessorOptions[];
+};
+export function buildDictionaryQueryEntries(
+  dataSources: GroupedEthereumProjectDs[],
+): DictionaryQueryEntry[] {
+  const queryEntries: DictionaryQueryEntry[] = [];
+
+  for (const ds of dataSources) {
+    for (const handler of ds.mapping.handlers) {
+      // No filters, cant use dictionary
+      if (!handler.filter) return [];
+
+      switch (handler.kind) {
+        case EthereumHandlerKind.Block:
+          return [];
+        case EthereumHandlerKind.Call: {
+          const filter = handler.filter as EthereumTransactionFilter;
+          if (
+            filter.from !== undefined ||
+            filter.to !== undefined ||
+            filter.function !== undefined
+          ) {
+            queryEntries.push(callFilterToQueryEntry(filter, ds.options));
+          } else {
+            return [];
+          }
+          break;
+        }
+        case EthereumHandlerKind.Event: {
+          const filter = handler.filter as EthereumLogFilter;
+          if (ds.groupedOptions) {
+            queryEntries.push(
+              eventFilterToQueryEntry(filter, ds.groupedOptions),
+            );
+          } else if (ds.options?.address || filter.topics) {
+            queryEntries.push(eventFilterToQueryEntry(filter, ds.options));
+          } else {
+            return [];
+          }
+          break;
+        }
+        default:
+      }
+    }
+  }
+
+  return uniqBy(
+    queryEntries,
+    (item) =>
+      `${item.entity}|${JSON.stringify(
+        sortBy(item.conditions, (c) => c.field),
+      )}`,
+  );
+}
 
 @Injectable()
 export class FetchService extends BaseFetchService<